+++ src/lib/viz/dataview/mode/DataViewLocal.ts
@@ -15,29 +15,31 @@ export class DataViewLocal extends DataViewMode {
     this.bindEvents();
   }
 
-  public getSourceData(columns: string[] = [], options: { filterId?: string } = {}) {
+  public async getSourceData(columns: string[] = [], options: { excludedFilters?: string[] } = {}) {
     if (!columns.includes(this.column)) {
       columns.push(this.column);
     }
 
+    const { excludedFilters = [] } = options;
+
     if (this.useViewport) {
-      const filterOptions = options.filterId ? [options.filterId] : [];
-      return (this.dataSource as Layer).getViewportFeatures(filterOptions);
+      await (this.dataSource as Layer).addSourceField(this.column);
+      return (this.dataSource as Layer).getViewportFeatures(excludedFilters);
     }
 
     // is GeoJSON Layer
     if (this.dataSource instanceof Layer) {
-      return (this.dataSource.source as GeoJSONSource).getFeatures(columns);
+      return (this.dataSource.source as GeoJSONSource).getFeatures(excludedFilters);
     }
 
     // is GeoJSON Source
-    return (this.dataSource as GeoJSONSource).getFeatures(columns);
+    return (this.dataSource as GeoJSONSource).getFeatures(excludedFilters);
   }
 
   public async groupBy(
     operationColumn: string,
     operation: AggregationType,
-    options: { filterId?: string }
+    options: { excludedFilters: string[] }
   ) {
     const sourceData = await this.getSourceData([operationColumn || this.column], options);
     const { groups, nullCount } = groupValuesByColumn(