+++ CHANGELOG.md
@@ -7,12 +7,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [Unreleased]
 ### Added
 - Add `beforeLayerId` and `afterLayerId` options to `carto.viz.Layer.addTo` method, to customize layer position
+- Widget synchronization when using remote dataviews ([#76](https://github.com/CartoDB/web-sdk/pull/76/))
 
 ### Changed
 - Remove d3.format option from `carto.viz.Popup`
 
 ### Fixed
-- 
+- Fix popup interactivity glitch on hover
 
 ## [1.0.0-alpha.0] 2020-07-06
 - First release